Rename rule to no-is-function
