+++ README.md
@@ -44,12 +44,12 @@ Add `jquery` to the plugins section of your `.eslintrc` configuration file. You 
     "jquery/no-fade": 2,
     "jquery/no-filter": 2,
     "jquery/no-find": 2,
-    "jquery/no-function": 2,
     "jquery/no-global-eval": 2,
     "jquery/no-has": 2,
     "jquery/no-hide": 2,
     "jquery/no-html": 2,
     "jquery/no-in-array": 2,
+    "jquery/no-is-function": 2,
     "jquery/no-is": 2,
     "jquery/no-load": 2,
     "jquery/no-map": 2,
