+++ index.js
@@ -18,12 +18,12 @@ module.exports = {
     'no-fade': require('./rules/no-fade'),
     'no-filter': require('./rules/no-filter'),
     'no-find': require('./rules/no-find'),
-    'no-function': require('./rules/no-function'),
     'no-global-eval': require('./rules/no-global-eval'),
     'no-has': require('./rules/no-has'),
     'no-hide': require('./rules/no-hide'),
     'no-html': require('./rules/no-html'),
     'no-in-array': require('./rules/no-in-array'),
+    'no-is-function': require('./rules/no-is-function'),
     'no-is': require('./rules/no-is'),
     'no-load': require('./rules/no-load'),
     'no-map': require('./rules/no-map'),
