+++ rules/no-is-function.js
